Add NDPluginSSCPimega to common plugins
SSCPimega plugin enables geometric restauration for Pimega detectors.
This is added as a default plugin to Pimega detectors, so that they can
be easily enabled in specific beamline configuration.

This is not a iocsh script configuration due to library linkage during
the build. Thus, this requires a patch to the IOC in `iocs` directory
from ADPimega.