Remove RAD mentions and pydm screenshots
As of now, ADPimega was only tested for PIMEGA models. Before adding
RAD to the docs we should make sure ADPimega is compatible with this
model.

The screenshots are from a custom GUI that only works with ADPimega.
We should add support for our IOC in the popular community GUIs like
MEDM and ImageJ.