Handle high perf stream data for 12 and 24 bit images
The high perf stream sends 16-bit data for 12-bit images, and 32-bit
data for 24-bit images, so updateEpicsFrames needed to be updated to
deal with that.

Accessing PimegaCounterDepth in the zmq callback without locking the
port is safe, because the value can't be updated while an acquisition is
happening, which is when the frames are received.

It's important to note that this change requires changing the
BackendPortFrame argument to pimegaDetectorConfig to match "stream_port"
from backend.json. This breaks the DACCountScan implementation, because
it depends on the port being passed to it having the
"pimega_frame_visualizer" subscription available. This issue isn't dealt
with here, because the DACCountScan implementation in the IOC is already
broken for other reasons.

Co-authored-by: Érico Nogueira <erico.rolim@lnls.br>