Don't share `pimega->error` buffer with ADPimega
Commit 8df56f3 added a `pimegaDetector::error_str` field that was set to
point to pimega-api's internal error buffer. This effectively shares the
api buffer with ADPimega, which is very error prone. Also, 8df56f3
writes to the buffer, possibly overwriting messages from pimega-api.

The present commit removes `pimegaDetector::error_str` and local `const
char*` to point either to `pimega->error` or to ADPimega's error
messages.

The use of a pointer to a constant array of chars gives the intended
semantics of not allowing someone to indirectly write to `pimega->error`
buffer.

Fixes: 8df56f3 (Replaced pimega->error and module/sensor/mb operator, 2024-02-26)